requirements.txt
paho-mqtt==2.1.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
